backend/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.3
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
flake8==6.0.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
